Moved definitions to a header file "src/header.h"
